Move graph interfaces into internal.interfaces
