+++ neuro_san/internals/graph/activations/abstract_callable_activation.py
@@ -20,7 +20,7 @@ from typing import List
 
 from langchain_core.messages.base import BaseMessage
 
-from neuro_san.internals.graph.interfaces.agent_tool_factory import AgentToolFactory
+from neuro_san.internals.interfaces.agent_tool_factory import AgentToolFactory
 from neuro_san.internals.interfaces.callable_activation import CallableActivation
 from neuro_san.internals.run_context.interfaces.agent_network_inspector import AgentNetworkInspector
 from neuro_san.internals.run_context.interfaces.run_context import RunContext
